Cleanup and minor refactoring.
