+++ mentions/forms/manual_submit_webmention.py
@@ -1,6 +1,10 @@
 from django import forms
 
+__all__ = [
+    "ManualSubmitWebmentionForm",
+]
+
 
 class ManualSubmitWebmentionForm(forms.Form):
-    target = forms.URLField(label='The URL of my page')
-    source = forms.URLField(label='The URL of your page')
+    target = forms.URLField(label="The URL of my page")
+    source = forms.URLField(label="The URL of your page")
