+++ mentions/tasks/incoming/remote.py
@@ -4,9 +4,17 @@ from typing import Optional
 from mentions.exceptions import SourceDoesNotLink, SourceNotAccessible
 from mentions.models import HCard
 from mentions.models.mixins.quotable import IncomingMentionType
-from mentions.tasks.incoming.parsing.hcard import find_related_hcard, parse_hcard
-from mentions.tasks.incoming.parsing.post_type import parse_post_type
+from mentions.tasks.incoming.parsing import (
+    find_related_hcard,
+    parse_hcard,
+    parse_post_type,
+)
 from mentions.util import html_parser, http_get
+
+__all__ = [
+    "get_source_html",
+    "get_metadata_from_source",
+]
 
 
 def get_source_html(source_url: str) -> str:
