+++ mentions/views/__init__.py
@@ -1,0 +1,3 @@
+from .dashboard import WebmentionDashboardView
+from .retrieve_mentions import GetMentionsView
+from .submit_mentions import WebmentionView
